requirements.txt
torch
torchaudio
numpy
soundfile
scikit-learn
speechbrain
pyserial
sounddevice
